API-3: :white_check_mark: test: make sure that only owner can delete

+++ app/Http/Controllers/Question/DeleteController.php
@@ -13,6 +13,8 @@ class DeleteController extends Controller
      */
     public function __invoke(Question $question)
     {
+        $this->authorize('forceDelete', $question);
+
         $question->forceDelete();
 
         return response()->noContent();
